Updt

+++ GitRes.bat
@@ -2,15 +2,22 @@
 REM Check if the argument %1 is provided
 
 IF "%~1"=="" (
-    echo Rerun code with commit description
+    echo Setting description to Updt
+    set Desc=Updt
 ) ELSE (
+  set Desc=%1
+)
+
+rem IF "%~1"=="" (
+rem    echo Rerun code with commit description
+rem ) ELSE (
     cd D:\Resumes
     set GIT_TRACE=1
     echo FORCE A RE-READING OF THE .gitignore list
     git rm -r --cached .
     git add .
     git branch -M main
-    git commit -m %1
+    git commit -m %Desc%
     REM COMMENTED OUT CODE FORCES A PUSH
     REM It loses track of remote updates
     REM git push origin main --force
@@ -20,4 +27,4 @@ IF "%~1"=="" (
     echo .
     echo VIEWS IF BATCH SUCCEEDED
     git log -n 1
-)
+rem )
